Rename DCM.py to dcm.py
In accordance with pep8: https://peps.python.org/pep-0008/#package-and-module-names
