+++ src/dodal/devices/undulator_dcm.py
@@ -3,7 +3,7 @@ from numpy import argmin, loadtxt, ndarray
 from ophyd import Component, Device, Signal
 from ophyd.status import Status
 
-from dodal.devices.DCM import DCM
+from dodal.devices.dcm import DCM
 from dodal.devices.undulator import Undulator, UndulatorGapAccess
 from dodal.log import LOGGER
